// server/src/reputation/reputation.enum.ts
export enum ReputationEvent {
  POST_CREATED = 'post_created',
  POST_LIKED = 'post_liked',
  REPOST = 'repost',
  COMMENT_CREATED = 'comment_created',
  TIP_SENT = 'tip_sent',
  TIP_RECEIVED = 'tip_received',
  PROPOSAL_CREATED = 'proposal_created',
  VOTE_CAST = 'vote_cast',
  DATA_SHARED = 'data_shared',
  INTERACTION_RECEIVED = 'interaction_received',
}